[Refactor] folderCell 분리
