+++ Cote/Cote/Presentation/Shared/Cell/Note+FolderCell.swift
@@ -166,29 +166,6 @@ struct ListCell: View {
     }
 }
 
-struct FolderCell: View {
-    @EnvironmentObject private var viewModel: SideBarViewModel
-    
-    let isExpanded: Bool
-    let folder: Folder
-    @Binding var isHover: Bool
-    
-    var body: some View {
-        HStack(spacing: 0) {
-            if isExpanded {
-                Image("arrow_down")
-                    .foregroundStyle(.iconSecondary)
-            } else {
-                Image("arrow_right")
-                    .foregroundStyle(.iconSecondary)
-            }
-            Text(folder.name)
-                .coteFont(.text2,
-                          color: isHover ? .textStrong : .textDefault)
-        }
-    }
-}
-
 extension Notification.Name {
     static let moveNoteRequest = Notification.Name("MoveNoteRequest")
 }
